+++ JayYi/ArchitectureStudy/app/src/main/java/com/jay/architecturestudy/ui/blog/BlogFragment.kt
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import com.jay.architecturestudy.R
 import com.jay.architecturestudy.data.model.Blog
 import com.jay.architecturestudy.data.model.ResponseNaverQuery
+import com.jay.architecturestudy.data.repository.NaverSearchRepositoryImpl
 import com.jay.architecturestudy.network.Api
 import com.jay.architecturestudy.ui.BaseFragment
 import kotlinx.android.synthetic.main.fragemnt_movie.*
@@ -16,6 +17,10 @@ import retrofit2.Response
 class BlogFragment : BaseFragment(R.layout.fragemnt_blog) {
 
     private lateinit var blogAdapter: BlogAdapter
+
+    private val naverSearchRepository by lazy {
+        NaverSearchRepositoryImpl()
+    }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
@@ -38,22 +43,14 @@ class BlogFragment : BaseFragment(R.layout.fragemnt_blog) {
     }
 
     override fun search(keyword: String) {
-        Api.getBlog(keyword)
-            .enqueue(object : Callback<ResponseNaverQuery<Blog>> {
-                override fun onFailure(call: Call<ResponseNaverQuery<Blog>>, t: Throwable) {
-                    Log.e("Blog", "error=${t.message}")
-                }
-
-                override fun onResponse(
-                    call: Call<ResponseNaverQuery<Blog>>,
-                    response: Response<ResponseNaverQuery<Blog>>
-                ) {
-                    if (response.isSuccessful) {
-                        val body = response.body() ?: return
-                        blogAdapter.setData(body.items)
-                    }
-                }
-
-            })
+        naverSearchRepository.getBlog(
+            keyword = keyword,
+            success = { responseBlog ->
+                blogAdapter.setData(responseBlog.blogs)
+            },
+            fail = { e ->
+                Log.e("Blog", "error=${e.message}")
+            }
+        )
     }
 }
